Remove unused UDPSocket::constructSocket(). Make the listeningPort an std::optional<unsigned int> to make it clear when the socket is not actually able to listen or receive. Update tests for listening port change.

+++ src/socket/Socket.cpp
@@ -56,7 +56,7 @@ namespace kt
 			timeOutVal = &timeoutVal;
 		}
 		
-		timeOutVal->tv_usec = static_cast<long>(timeout);
+		timeOutVal->tv_usec = timeout;
 
 		FD_ZERO(&sReady);
 		FD_SET(socketDescriptor, &sReady);
